pyproject.toml
[project]
name = "bootdev-aiagent"
version = "0.1.0"
description = "Buidling Python AI Agents with Gemini"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
